requirements.txt
openai==1.86.0
langchain==0.3.25
langchain-openai==0.3.22
google-ai-generativelanguage==0.6.18
langchain-google-genai==2.1.8
langgraph==0.4.8
python-dotenv==1.0.1
langchainhub==0.1.15
pandas==2.2.2
faiss-cpu==1.11.0
langchain-community==0.3.25
pypdf==5.6.0